feat: :sparkles: feat: ambulance Add Addition operator
the addition operator (+) is a binary operator that is used to perform arithmetic addition or string concatenation, depending on the types of the operands. When used with two numbers, it adds them together to produce a numeric result. When used with a string and a non-string value, it performs implicit type coercion and concatenates the values into a new string.